requirements.txt
flask==2.2.3
python-dotenv==1.0.0
uuid==1.30
pathlib==1.0.1